requirements.txt
asgiref==3.5.0
cssselect2==0.4.1
Django==3.0
django-cleanup==6.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
easy-thumbnails==2.8.1
lxml==4.7.1
Pillow==9.0.1
psycopg2==2.8
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
webencodings==0.5.1
